+++ TemplateRESTful.Service/Server/Interfaces/IEmailService.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using TemplateRESTful.Infrastructure.Server.Requests;
+
+namespace TemplateRESTful.Service.Server.Interfaces
+{
+    public interface IEmailService
+    {
+        Task SendEmailAsync(EmailMessage emailMessage);
+    }
+}