+++ TemplateRESTful.Web/Areas/Accounts/Controllers/ProfilesController.cs
@@ -1,13 +1,14 @@
-﻿using System.Threading.Tasks;
+﻿using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
-using TemplateRESTful.Domain.Models.DTOs;
 using TemplateRESTful.Domain.Models.Entities;
-using TemplateRESTful.Domain.Models.Entities.Profiles;
-using TemplateRESTful.Persistence.Data.Profiles.Commands;
-using TemplateRESTful.Persistence.Data.Profiles.Queries;
+using TemplateRESTful.Service.Client.Actions.Commands;
+using TemplateRESTful.Service.Client.Actions.Queries;
+using TemplateRESTful.Service.Client.Interfaces;
 using TemplateRESTful.Web.Areas.Accounts.Models;
 using TemplateRESTful.Web.Controllers;
 
@@ -17,24 +18,28 @@ namespace TemplateRESTful.Web.Areas.Accounts.Controllers
     public class ProfilesController : RootController<ProfilesController>
     {
         private readonly UserManager<ApplicationUser> _userManager;
+        private readonly IOnlineProfileService _profileManager;
 
-        public ProfilesController(UserManager<ApplicationUser> userManager)
+        public ProfilesController(UserManager<ApplicationUser> userManager,
+            IOnlineProfileService profileManager)
         {
             _userManager = userManager;
+            _profileManager = profileManager;
         }
 
         [HttpGet]
-        public async Task<IActionResult> UpdateProfile(int profileId = 0)
+        public async Task<IActionResult> SaveProfile(string profileId = null)
         {
             var onlineUser = await _userManager.GetUserAsync(User);
-            var onlineProfile = LoadUserProfile(onlineUser);
+            var onlineProfile = _profileManager.SetOnlineProfile(onlineUser);
 
-            if (profileId != 0) 
+            if (profileId != null) 
             {
                 var existingProfile = await _mediator.Send(
                     new GetOnlineProfileQuery()
                     {
                         Id = profileId,
+                        EmailAddress = onlineProfile.EmailAddress
                     }
                 );
 
@@ -65,17 +70,29 @@ namespace TemplateRESTful.Web.Areas.Accounts.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> UpdateProfile(int profileId,
+        public async Task<IActionResult> SaveProfile(string profileId,
             UserProfileViewModel profileModel, string returnUrl = null)
         {
+            var currentProfile = await _profileManager.GetByIdAsync(profileId);
             var onlineProfile = profileModel.OnlineProfile;
+
             returnUrl ??= Url.Content("~/");
 
             if (ModelState.IsValid)
             {
-                if (profileId > 0)
+                if (profileId != null)
                 {
                     onlineProfile.Id = profileId;
+
+                    if (Request.Form.Files.Count > 0)
+                    {
+                        await ChangeProfilePicture(onlineProfile);
+                    }
+                    else
+                    {
+                        onlineProfile.Picture = currentProfile.Picture;
+                    }
+
                     var updateProfile = _mapper.Map<UpdateProfileCommand>(onlineProfile);
                     await _mediator.Send(updateProfile);
 
@@ -83,6 +100,11 @@ namespace TemplateRESTful.Web.Areas.Accounts.Controllers
                 }
                 else
                 {
+                    if (Request.Form.Files.Count > 0)
+                    {
+                        await ChangeProfilePicture(onlineProfile);
+                    }
+
                     var createProfile = _mapper.Map<CreateProfileCommand>(onlineProfile);
                     await _mediator.Send(createProfile);
 
@@ -93,24 +115,20 @@ namespace TemplateRESTful.Web.Areas.Accounts.Controllers
             return View(profileModel);
         }
 
-        private OnlineProfile LoadUserProfile(ApplicationUser onlineUser)
+        private async Task<byte[]> ChangeProfilePicture(OnlineProfile profile)
         {
-            var profileUser = new OnlineProfileDto();
-            var profileModel = new OnlineProfile
+            if (Request.Form.Files.Count > 0)
             {
-                FirstName = onlineUser.FirstName,
-                MiddleName = profileUser.MiddleName,
-                LastName = onlineUser.LastName,
-                PhoneNumber = onlineUser.PhoneNumber,
-                EmailAddress = onlineUser.Email,
-                DayOfBirth = profileUser.DayOfBirth,
-                Occupation = profileUser.Occupation,
-                Website = profileUser.Website,
-                Location = profileUser.Location,
-                Language = profileUser.Language
-            };
+                IFormFile picture = Request.Form.Files.FirstOrDefault();
 
-            return profileModel;
+                using (var dataStream = new MemoryStream())
+                {
+                    await picture.CopyToAsync(dataStream);
+                    profile.Picture = dataStream.ToArray();
+                }
+            }
+
+            return default;
         }
     }
 }