+++ TemplateRESTful.Web/Resources/SharedResource.cs
@@ -1,5 +1,0 @@
-﻿
-namespace TemplateRESTful.Web.Resources
-{
-    public class SharedResource {}
-}