+++ TemplateRESTful.Web/Views/ViewModels/NotyfNotificationVM.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-using TemplateRESTful.Service.Features.Notifications;
+using TemplateRESTful.Service.Client.Features;
 
 namespace TemplateRESTful.Web.Shared.ViewModels
 {
     public class NotyfNotificationVM
     {
         public string Configuration { get; set; }
-        public IEnumerable<NotyfNotification> NotyfNotifications { get; set; }
+        public IEnumerable<NotyfNotification> Notifications { get; set; }
     }
 }